feat: enhance interactive predictions and add main script for training and evaluation
